+++ src/Toolkit/CHANGELOG.md
@@ -12,7 +12,12 @@
 - [Shadcn] Rework templates of `input-group` recipe
 - [Shadcn] Rework templates of `table` recipe
 - [Shadcn] Rework templates of `textarea` recipe
+- [Shadcn] Add `Avatar:Badge` component
+- [Shadcn] Add `Avatar:Fallback` component
+- [Shadcn] Add `Avatar:Group` component
+- [Shadcn] Add `Avatar:GroupCount` component
 - [Shadcn] Add `Alert:Action` component
+- [Shadcn] Remove `Avatar:Text` component
 
 ## 2.29.0